Split Main class into TwitterClient; removed from solutions
