+++ projects/Pitovsky/pom.xml
@@ -9,7 +9,7 @@
   </parent>
   <groupId>ru.mipt.diht.students</groupId>
   <artifactId>Pitovsky</artifactId>
-  <version>1.1-SNAPSHOT</version>
+  <version>1.2-SNAPSHOT</version>
   <name>Pitovsky</name>
   <url>http://maven.apache.org</url>
   <dependencies>
