--- solutions/pitovsky/TwitterStream/pom.xml
@@ -1,65 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>Pitovsky</groupId>
-  <artifactId>TwitterStream</artifactId>
-  <version>0.1.0-SNAPSHOT</version>
-  
-  <parent>
-    <groupId>ru.mipt.diht.students</groupId>
-    <artifactId>parent</artifactId>
-    <version>1.0-SNAPSHOT</version>
-  </parent>
-  
-  <dependencies>
-     <dependency>
-         <groupId>org.twitter4j</groupId>
-         <artifactId>twitter4j-core</artifactId>
-         <version>[4.0,)</version>
-     </dependency>
-     <dependency>
-         <groupId>org.twitter4j</groupId>
-         <artifactId>twitter4j-stream</artifactId>
-         <version>[4.0,)</version>
-     </dependency>
-     <dependency>
-		 <groupId>com.beust</groupId>
-		 <artifactId>jcommander</artifactId>
-		 <version>1.48</version>
-	 </dependency>
-  </dependencies>
-    
-  <build>
-    <sourceDirectory>src</sourceDirectory>
-    <plugins>
-      <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.1</version>
-        <configuration>
-          <source>1.7</source>
-          <target>1.7</target>
-        </configuration>
-      </plugin>
-      <plugin>
-       <artifactId>maven-assembly-plugin</artifactId>
-       <configuration>
-        <descriptorRefs>
-         <descriptorRef>jar-with-dependencies</descriptorRef>
-        </descriptorRefs>
-        <archive>
-         <manifest>
-           <mainClass>main.java.ru.fizteh.fivt.pitovsky.Main</mainClass>
-         </manifest>
-       </archive>
-      </configuration>
-      <executions>
-        <execution>
-          <phase>package</phase>
-          <goals>
-            <goal>single</goal>
-          </goals>
-        </execution>
-      </executions>
-     </plugin>
-    </plugins>
-  </build>
-</project>
